0.1.2 - Добавил инструмент рисования стрелки - Модифицировал рисование линии и стрелки для рисования чисто горизонтальных или чисто вертикальных линий. 	Их рисование выполняется с пом. нажатия Ctrl|Alt|Shift во время рисования. - Добавил опцию использовать только родные диалоги. - Поправил хранение настроек в Windows.

+++ sources/dialogs/effectsettingsdialog.cpp
@@ -67,7 +67,7 @@ EffectSettingsDialog::EffectSettingsDialog(const QImage &img, AbstractEffectSett
 
 QRgb EffectSettingsDialog::convolutePixel(const QImage &image, int x, int y, const QList<double> &kernelMatrix)
 {
-    int kernelSize = sqrt(kernelMatrix.size());
+    int kernelSize = sqrt(double(kernelMatrix.size()));
 
     double total = 0;
     double red = 0;
